Moving validation tests to a dedicated class
